requirements.txt
flask
flask-cors
streamlit
requests
pandas
numpy
plotly
python-dotenv
scikit-learn
tensorflow
uvicorn
fastapi
pydantic
